backend/requirements.txt
fastapi
uvicorn
pandas
yfinance
scipy
newspaper3k
ta
matplotlib
google-generativeai
pymupdf
lxml_html_clean
python-multipart
